Revert "Merge branch 'interface' of https://github.com/mooch443/trex into interface"
This reverts commit 0852cfdbd9c6097290b3b2644afab2be5851afb0, reversing
changes made to a31f8b5a16af901d33f1f32e9e20fa6ed8945ce9.

+++ Application/src/tracker/tracking/TrackingHelper.cpp
@@ -493,7 +493,7 @@ double TrackingHelper::process_postures() {
             }
             
             std::lock_guard guard((_statistics_mutex));
-            combined_posture_seconds += narrow_cast<float>(collected);
+            combined_posture_seconds += collected;
             
         }, Tracker::instance()->thread_pool(), all.begin(), all.end());
